Object Oriented PHP
Class and Object Refresher
Object-Oriented PHP helps your code stay flexible by allowing it to be only defined once but used in many places.

class ClassName {
}


$className = new ClassName();
We've already learned how to create objects from the classes we define, so let's run through a little refresher. Remember, we use the class keyword to define our classes.
<html>
  <head>
    <title>OOP Review</title>
  </head>
  <body>
    <p>
      <?php
      // Create a Person class here:
      class Person {
}

$me = new Person();
      
      // And create a Person instance called $me here:
      
      
      ?>
    </p>
  </body>
</html>

Class and Object Methods
Remember when we covered built-in functions in PHP? Well, we can combine that idea with our knowledge of classes and objects: that is, there are built-in PHP functions that tell us interesting information about the classes and objects we've created!

Check out the example in the editor to the right. We're demoing three useful built-in methods: is_a(), which we use to find out if a particular object is an instance of a given class; property_exists(), to see if an object has a given property; and method_exists(), to see if an object has a given method.
<html>
  <head>
    <title>Class and Object Methods</title>
  </head>
  <body>
    <p>
      <?php
        class Person {
          public $isAlive = true;
          
          function __construct($name) {
              $this->name = $name;
          }
          
          public function dance() {
            return "I'm dancing!";
          }
        }
        
        $me = new Person("Shane");
        if (is_a($me, "Person")) {
          echo "I'm a person, ";
        }
        if (property_exists($me, "name")) {
          echo "I have a name, ";
        }
        if (method_exists($me, "dance")) {
          echo "and I know how to dance!";
        }
      ?>
    </p>
  </body>
</html>
Note that the first argument is the object we're checking, and the second is the class, property, or method name as a "string".
INHARITANCE
As you've been thinking about classes and objects, you might have realized that one class might actually be a specialized type of another class. For instance, you might have a Vehicle class and a Truck class, and it would probably save you an awful lot of typing if you could somehow specify that Truck instances should automatically have many of the same properties and methods as Vehicle instances.

PHP allows us to accomplish this through a process called inheritance. Inheritance is a way for one class to take on the properties or methods of another class. You could say that the one class extends the other. This is used to express an "is-a" relationship—for example, a Truck "is-a" Vehicle, so it could inherit from Vehicle, but a Motorcycle isn't a Truck, so it shouldn't inherit from Truck (though both could inherit from Vehicle).

We can cause one PHP class to inherit from another with the extends keyword.
<html>
  <head>
    <title>The Shape of Things to Come</title>
  </head>
  <body>
    <p>
      <?php
        class Shape {
          public $hasSides = true;
        }
        
        class Square extends Shape {
        
        }
        
        $square = new Square();
      
        if ( property_exists($square, "hasSides")) {
          echo "I have sides!";
        }
      ?>
    </p>
  </body>
</html>

Overriding Parent Methods
Sometimes we want a child class (or subclass) to be able to override a property or method of its parent class (or superclass).

For instance, we might have a Shape class with a $sides property set to true, but we might want Square to override this property and set $sides to 4 (since a square always has four sides). That would look something like this:

class Shape {
  $sides = true;
}

class Square extends Shape {
  $sides = 4;
}
It's pretty easy—you just create a new property or method in the child class with the same name as the one in the parent class, and the child's version will always take precedence over the inherited version.
<html>
  <head>
    <title>Override!</title>
  </head>
  <body>
    <p>
   <?php
class Vehicle {
  public function honk() {
    return "HONK HONK!";
  }
}
// Add your code below!
class Bicycle extends Vehicle {
  public function honk() {
    return "Beep beep!";
  }
}
$bicycle = new Bicycle();
echo $bicycle->honk();
?>
    </p>
  </body>
</html>

The Static Keyword
You probably noticed that we could access the Ninja class constant without actually creating an instance of Ninja, and if you're particularly precocious, you might be wondering whether it's possible to access class properties or methods without creating an instance of the class. The answer: yes!

The static keyword lets you use a class' property or method without having to create an instance of that class. It works like this:

class Person {
  public static $isAlive = "Yep!"
  public static function greet() {
    echo "Hello there!";
  }
}

echo Person::$isAlive;
// prints "Yep!"
Person::greet();
// prints "Hello there!"
When combined with the scope resolution operator, this lets us access class information without having to instantiate anything.

<html>
  <head>
    <title></title>
  </head>
  <body>
    <p>
<?php
class King {
   public static $isAlive = "Yep!";
    
public  static function proclaim()
{
echo "A kingly proclamation!";
}
}
King::proclaim();
echo King::$isAlive;
?>
    </p>
  </body>
</html>

Class Constants and Scope Resolution
We've talked a lot about changing variables in PHP, but sometimes we want variables that don't change. These are prefixed with the const keyword (short for constant).
class Immortal extends Person {
  // Immortals never die!
  const alive = true;
}

// If true...
if (Immortal::alive) {
  echo "I live forever!";
}
PHP lets us set constants on a class-by-class basis! Each class has its own scope, which is the context in which its variables can be used.
 we use :: to access the alive constant inside the Immortal class.Note that constants do not start with $.
<html>
  <head>
    <title>Scope it Out!</title>
  </head>
  <body>
    <p>
   <?php
class Person {

  }
  class Ninja extends Person {
    // Add your code here...
    const stealth = "MAXIMUM";

  }
  // ...and here!

      echo Ninja::stealth;


  ?>
    </p>
  </body>
</html>